@@ -18,12 +18,12 @@
         }
     ],
     "require": {
-        "php": ">=5.5.9",
+        "php": ">=7.2.5",
         "doctrine/collections": "^1.3",
         "symfony/form": "^4.3|^5.0",
         "symfony/framework-bundle": "^4.3|^5.0",
-        "symfony/security": "^4.3 || ^5.0",
-        "symfony/twig-bundle": "^4.3|^5.0"
+        "symfony/twig-bundle": "^4.3|^5.0",
+        "symfony/asset": "^5.0"
     },
     "require-dev": {
         "symfony/validator": "^4.3|^5.0",
