@@ -13,11 +13,13 @@ interface DeleterInterface
 {
     /**
      * Marks the thread as deleted by the current authenticated user.
+     * @param ThreadInterface $thread
      */
-    public function markAsDeleted(ThreadInterface $thread);
+    public function markAsDeleted(ThreadInterface $thread): void;
 
     /**
      * Marks the thread as undeleted by the current authenticated user.
+     * @param ThreadInterface $thread
      */
-    public function markAsUndeleted(ThreadInterface $thread);
+    public function markAsUndeleted(ThreadInterface $thread): void;
 }
