@@ -5,6 +5,7 @@ namespace FOS\MessageBundle\FormHandler;
 use FOS\MessageBundle\FormModel\AbstractMessage;
 use FOS\MessageBundle\FormModel\NewThreadMessage;
 use FOS\MessageBundle\Model\MessageInterface;
+use InvalidArgumentException;
 
 class NewThreadMessageFormHandler extends AbstractMessageFormHandler
 {
@@ -13,14 +14,14 @@ class NewThreadMessageFormHandler extends AbstractMessageFormHandler
      *
      * @param AbstractMessage $message
      *
-     * @throws \InvalidArgumentException if the message is not a NewThreadMessage
+     * @throws InvalidArgumentException if the message is not a NewThreadMessage
      *
      * @return MessageInterface the composed message ready to be sent
      */
     public function composeMessage(AbstractMessage $message)
     {
         if (!$message instanceof NewThreadMessage) {
-            throw new \InvalidArgumentException(sprintf('Message must be a NewThreadMessage instance, "%s" given', get_class($message)));
+            throw new InvalidArgumentException(sprintf('Message must be a NewThreadMessage instance, "%s" given', get_class($message)));
         }
 
         return $this->composer->newThread()
