@@ -5,6 +5,7 @@ namespace FOS\MessageBundle\FormHandler;
 use FOS\MessageBundle\FormModel\AbstractMessage;
 use FOS\MessageBundle\FormModel\NewThreadMultipleMessage;
 use FOS\MessageBundle\Model\MessageInterface;
+use InvalidArgumentException;
 
 /**
  * Form handler for multiple recipients support.
@@ -18,14 +19,14 @@ class NewThreadMultipleMessageFormHandler extends AbstractMessageFormHandler
      *
      * @param AbstractMessage $message
      *
-     * @throws \InvalidArgumentException if the message is not a NewThreadMessage
+     * @throws InvalidArgumentException if the message is not a NewThreadMessage
      *
      * @return MessageInterface the composed message ready to be sent
      */
     public function composeMessage(AbstractMessage $message)
     {
         if (!$message instanceof NewThreadMultipleMessage) {
-            throw new \InvalidArgumentException(sprintf('Message must be a NewThreadMultipleMessage instance, "%s" given', get_class($message)));
+            throw new InvalidArgumentException(sprintf('Message must be a NewThreadMultipleMessage instance, "%s" given', get_class($message)));
         }
 
         return $this->composer->newThread()
