@@ -2,8 +2,9 @@
 
 namespace FOS\MessageBundle\FormType;
 
-use FOS\MessageBundle\Util\LegacyFormHelper;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\TextareaType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 
 /**
@@ -16,18 +17,18 @@ class NewThreadMultipleMessageFormType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('recipients', LegacyFormHelper::getType('FOS\MessageBundle\FormType\RecipientsType'), array(
+            ->add('recipients', RecipientsType::class, [
                 'label' => 'recipients',
                 'translation_domain' => 'FOSMessageBundle',
-            ))
-            ->add('subject', LegacyFormHelper::getType('Symfony\Component\Form\Extension\Core\Type\TextType'), array(
+            ])
+            ->add('subject', TextType::class, [
                 'label' => 'subject',
                 'translation_domain' => 'FOSMessageBundle',
-            ))
-            ->add('body', LegacyFormHelper::getType('Symfony\Component\Form\Extension\Core\Type\TextareaType'), array(
+            ])
+            ->add('body', TextareaType::class, [
                 'label' => 'body',
                 'translation_domain' => 'FOSMessageBundle',
-            ));
+            ]);
     }
 
     /**
