@@ -2,6 +2,8 @@
 
 namespace FOS\MessageBundle\Search;
 
+use Symfony\Component\HttpFoundation\Request;
+
 /**
  * Gets the search term from the request and prepares it.
  */
@@ -10,7 +12,9 @@ interface QueryFactoryInterface
     /**
      * Gets the search term.
      *
+     * @param Request $request
+     *
      * @return Query the term object
      */
-    public function createFromRequest();
+    public function createFromRequest(Request $request);
 }
