@@ -18,6 +18,6 @@ class ReplyAuthorization extends Constraint
      */
     public function getTargets()
     {
-        return self::CLASS_CONSTRAINT;
+        return static::CLASS_CONSTRAINT;
     }
 }
